requirements.txt
streamlit==1.39.0
google-genai>=1.0.0,<2.0.0
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
sentence-transformers==3.2.1
chromadb==0.5.15
langgraph==0.2.45
langchain-core==0.3.15
transformers==4.45.2
torch==2.4.1
python-dotenv==1.0.1
